A large set of edits aimed at making the system more compatible with full 64-bit mode, such as avoiding truncation of 64-bit  offsets, etc.  Fixed a bunch of Makefiles which did not pass down the arch flags to their link commands, which caused many executables to fail to link.

+++ genlib/Odelay.cpp
@@ -101,7 +101,7 @@ static float *newFloats(float *oldptr, long oldlen, long *newlen)
 	return ptr;
 }
 
-int Odelay::resize(long thisLength)
+long Odelay::resize(long thisLength)
 {
 	const long oldlen = _len;
 	// Make a guess at how big the new array should be.
